refactor: change GifflarContractManager to GifflarManager
