--- src/modules/managing/contractManager/index.ts
@@ -1,12 +1,0 @@
-import { container } from "tsyringe";
-import GifflarContractManager from "./implementations/GifflarContractManager";
-import { IContractManager } from "./types/IContractManager";
-
-const implementations = {
-  default: GifflarContractManager,
-};
-
-container.registerSingleton<IContractManager>(
-  "GifflarContractManager",
-  implementations.default
-);
