+++ src/modules/managing/gifflarContract/types/IGifflarContract.ts
