+++ src/modules/managing/gifflarContract/types/IGifflarContractModel.ts
