+++ src/modules/managing/gifflarManager/types/IManagerDeployDTO.ts
